HTML de landing page, algunos íconos, archivo CSS principal, y reset CSS
