+++ frontend/src/app/app.routes.ts
@@ -1,3 +1,6 @@
 import { Routes } from '@angular/router';
+import { HomeComponent } from './screens/home/home.component';
 
-export const routes: Routes = [];
+export const routes: Routes = [
+  {path: '', component: HomeComponent},
+];
