+++ frontend/src/app/components/app/app.component.html
@@ -1,0 +1,6 @@
+<app-navbar />
+<main>
+  <router-outlet />
+</main>
+
+<app-footer />
