+++ frontend/src/app/components/footer/footer.component.html
@@ -1,0 +1,5 @@
+<footer>
+  <a href="#">
+    Hecho con 💖 - <strong>AngelMaldonado</strong>
+  </a>
+</footer>
