+++ frontend/src/app/components/navbar/navbar.component.html
@@ -1,0 +1,11 @@
+<nav>
+  <img src="logo.svg">
+  <ul>
+    <li><a href="#">Inicio</a></li>
+    <li><a href="#">Sobre mí</a></li>
+    <li><a href="#">Habilidades</a></li>
+    <li><a href="#">Proyectos</a></li>
+    <li><a href="#">Contacto</a></li>
+    <li><a href="#">Blog</a></li>
+  </ul>
+</nav>
